+++ packages/api-v4/src/account/index.ts
@@ -11,5 +11,3 @@ export * from './users';
 export * from './oauth';
 
 export * from './types';
-
-export * from './account.schema';
