+++ packages/api-v4/src/images/index.ts
@@ -1,5 +1,3 @@
 export * from './types';
 
-export * from './images.schema';
-
 export * from './images';
